+++ lib/widgets/ui/text/subtitle.dart
@@ -15,6 +15,7 @@ class SubTitle extends StatelessWidget {
         style: TextStyle(
           fontSize: 16.0,
           fontWeight: FontWeight.normal,
+          fontFamily: 'Heebo',
           color: Colors.white70
         ),
       ),